refactored Quicksort into Sorter interface
