+++ SortingMethods/Sorter.h
@@ -1,6 +1,17 @@
 #pragma once
-class Sorter<type T> {
+template<typename T>
+class Sorter {
 public:
-	virtual void sort(T[] ar) = 0;
 
+    Sorter() {}
+    virtual ~Sorter() {}
+
+    virtual void sort(T ar[], int low, int high) = 0;
+
+protected:
+    void swap(T* a, T* b) {
+        T t = *a;
+        *a = *b;
+        *b = t;
+    }
 };
