+++ SortingMethods/SortingMethods.vcxproj.filters
@@ -23,5 +23,8 @@
     <ClInclude Include="Sorter.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="Quicksort.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
